+++ platforms/fabric-1.20/src/main/java/org/anti_ad/mc/ipnext/item/ItemTypeExtensions.kt
@@ -236,9 +236,13 @@ private fun initGroupIndex(): Boolean {
 val ItemType.searchTabIndex: Int
     get() {
         initGroupIndex()
-        return ItemGroups.getSearchGroup().searchTabStacks.indexOfFirst { it ->
-            it != null && it.`(itemType)` == this
+        val index = ItemGroups.getSearchGroup().searchTabStacks.indexOfFirst { it ->
+            it != null && it.`(itemType)`.item == this.item
         }
+        if (index == -1) {
+            Log.warn("Item ${this.itemId} not present in the Creative Inventory Search Tab!")
+        }
+        return index
     }
 
 val ItemType.groupIndex: Int